// apps/desktop/src/generated/EventEnvelope.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CorrelationId } from "./CorrelationId";

/**
 * Versioned, correlated event envelope.
 */
export type EventEnvelope<T> = {
  schema_version: number;
  correlation_id: CorrelationId;
  payload: T;
};