// apps/desktop/src/generated/ListModulesResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EffectiveModule } from "./EffectiveModule";
import type { ModuleId } from "./ModuleId";

export type ListModulesResponse = {
  modules: Array<EffectiveModule>;
  unknown_preference_ids: Array<ModuleId>;
};